+++ src/llama_inference/__init__.py
@@ -1,6 +1,6 @@
 """Inference API for LLaMA"""
 
-from .app import ServeLLaMA
+from .serve import ServeLLaMA
 from .model import LLaMAInference
 
-__version__ = "0.0.1"
+__version__ = "0.0.2"
